Add lan module to allow switching off of lan adapter.
For some reason, the secondary lan port on the router stops working when the internet connection dies (my guess is that  router is testing if the remaining lan port can be used for the wan connection). This leads the device to not be reachable and unable to communicate out over the lan connection. Disabling the connection allows the allows the wifi connection to be used for managing the router instead.